cms-admin: Restructure action log module into three top-level folders
Reshape src/actionLog into three first-level folders that mirror the
three product use cases:
- components/   shared dumb pieces used by all views
- actionLog/    entity-scoped per-record components (Button, Dialog,
                VersionGrid) - the existing actionLogGrid is renamed to
                actionLogVersionGrid since each row represents a version
                of a single entity instance
- globalActionLog/  cross-scope activity feed components (renamed from
                actionLogs/ to avoid the ActionLog vs ActionLogs typo
                trap)

Symbol renames:
- ActionLogGrid              -> ActionLogVersionGrid
- actionLogGridFragment      -> actionLogVersionGridFragment
- ActionLogsGrid             -> GlobalActionLogGrid
- ActionLogsPage             -> GlobalActionLogPage
- ActionLogsDialog           -> GlobalActionLogDialog
- ActionLogsShowVersionDialog -> GlobalActionLogShowVersionDialog
- actionLogsGridFragment     -> globalActionLogGridFragment

Translation keys, GraphQL operation names, persistent-column-state
keys and Storybook titles all updated to match. Demo MasterMenu and
Storybook mock handlers updated.
