+++ packages/admin/cms-admin/src/actionLog/actionLog/actionLogVersionGrid/ActionLogVersionGrid.gql.ts
@@ -1,0 +1,14 @@
+import { gql } from "@apollo/client";
+
+export const actionLogVersionGridFragment = gql`
+    fragment ActionLogVersionGrid on ActionLog {
+        id
+        user {
+            id
+            name
+        }
+        entityName
+        version
+        createdAt
+    }
+`;
