--- packages/admin/cms-admin/src/actionLog/actionLogs/actionLogsDialog/__stories__/ActionLogsDialog.stories.tsx
@@ -1,27 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react-vite";
-
-import { ActionLogsDialog } from "../ActionLogsDialog";
-
-type StoryArgs = {
-    entityName: string;
-    entityId: string;
-    open: boolean;
-    onClose: () => void;
-};
-
-type Story = StoryObj<StoryArgs>;
-
-const meta: Meta<StoryArgs> = {
-    component: ActionLogsDialog,
-    tags: ["!autodocs"],
-    title: "actionLog/actionLogs/actionLogsDialog/ActionLogsDialog",
-    args: {
-        entityName: "News",
-        entityId: "550e8400-e29b-41d4-a716-446655440003",
-        open: true,
-        onClose: () => undefined,
-    },
-};
-export default meta;
-
-export const Default: Story = {};
