--- packages/admin/cms-admin/src/actionLog/actionLogs/actionLogsGrid/__stories__/ActionLogsGrid.stories.tsx
@@ -1,28 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react-vite";
-
-import { ContentScopeProvider } from "../../../../contentScope/Provider";
-import { ActionLogsGrid } from "../ActionLogsGrid";
-
-type Story = StoryObj<typeof ActionLogsGrid>;
-
-const meta: Meta<typeof ActionLogsGrid> = {
-    component: ActionLogsGrid,
-    tags: ["!autodocs"],
-    title: "actionLog/actionLogs/actionLogsGrid/ActionLogsGrid",
-    decorators: [
-        (Story) => (
-            <ContentScopeProvider
-                values={[
-                    { scope: { domain: "main", language: "en" }, label: { domain: "Main", language: "English" } },
-                    { scope: { domain: "main", language: "de" }, label: { domain: "Main", language: "German" } },
-                ]}
-                defaultValue={{ domain: "main", language: "en" }}
-            >
-                {() => <Story />}
-            </ContentScopeProvider>
-        ),
-    ],
-};
-export default meta;
-
-export const Default: Story = {};
