--- packages/admin/cms-admin/src/actionLog/actionLogs/actionLogsShowVersionDialog/ActionLogsShowVersionDialog.gql.ts
@@ -1,15 +1,0 @@
-import { gql } from "@apollo/client";
-
-import { actionLogCompareFragment } from "../../components/actionLogCompare/ActionLogCompare";
-
-export const actionLogsShowVersionQuery = gql`
-    query ActionLogsShowVersion($id: ID!) {
-        actionLog(id: $id) {
-            ...ActionLogCompare
-            previousVersion {
-                ...ActionLogCompare
-            }
-        }
-    }
-    ${actionLogCompareFragment}
-`;
