+++ packages/admin/cms-admin/src/actionLog/globalActionLog/globalActionLogShowVersionDialog/GlobalActionLogShowVersionDialog.gql.ts
@@ -1,0 +1,15 @@
+import { gql } from "@apollo/client";
+
+import { actionLogCompareFragment } from "../../components/actionLogCompare/ActionLogCompare";
+
+export const globalActionLogShowVersionQuery = gql`
+    query GlobalActionLogShowVersion($id: ID!) {
+        actionLog(id: $id) {
+            ...ActionLogCompare
+            previousVersion {
+                ...ActionLogCompare
+            }
+        }
+    }
+    ${actionLogCompareFragment}
+`;
