+++ packages/admin/cms-admin/src/index.ts
@@ -1,13 +1,13 @@
-export { ActionLogButton } from "./actionLog/actionLogButton/ActionLogButton";
-export { ActionLogDialog } from "./actionLog/actionLogDialog/ActionLogDialog";
-export { ActionLogGrid, actionLogGridFragment } from "./actionLog/actionLogGrid/ActionLogGrid";
-export { ActionLogsGrid, actionLogsGridFragment } from "./actionLog/actionLogs/actionLogsGrid/ActionLogsGrid";
-export { ActionLogsPage } from "./actionLog/actionLogs/actionLogsPage/ActionLogsPage";
+export { ActionLogButton } from "./actionLog/actionLog/actionLogButton/ActionLogButton";
+export { ActionLogDialog } from "./actionLog/actionLog/actionLogDialog/ActionLogDialog";
+export { ActionLogVersionGrid, actionLogVersionGridFragment } from "./actionLog/actionLog/actionLogVersionGrid/ActionLogVersionGrid";
 export { ActionLogCompare, actionLogCompareFragment } from "./actionLog/components/actionLogCompare/ActionLogCompare";
 export { ActionLogHeader, type ActionLogHeaderProps } from "./actionLog/components/actionLogHeader/ActionLogHeader";
 export { ActionLogShowVersion, actionLogShowVersionFragment } from "./actionLog/components/actionLogShowVersion/ActionLogShowVersion";
 export { DiffHeader, type DiffHeaderProps } from "./actionLog/components/diffHeader/DiffHeader";
 export { DiffViewer, type DiffViewerProps } from "./actionLog/components/diffViewer/DiffViewer";
+export { GlobalActionLogGrid, globalActionLogGridFragment } from "./actionLog/globalActionLog/globalActionLogGrid/GlobalActionLogGrid";
+export { GlobalActionLogPage } from "./actionLog/globalActionLog/globalActionLogPage/GlobalActionLogPage";
 export { AnchorBlock } from "./blocks/AnchorBlock";
 export { BlockAdminComponentButton } from "./blocks/common/BlockAdminComponentButton";
 export { BlockAdminComponentNestedButton } from "./blocks/common/BlockAdminComponentNestedButton";
